Respect the extension flag in linalg and fft
This behavior still needs to be tested.

This required moving the linalg functions that are also in the main namespace
so that they can still work there even when the linalg extension is disabled.

The way I've decided to implement this is that the functions will not raise an
exception until they are called. It would probably be more convenient for
users if they raised an attribute error, or if the extension namespace itself
did, like it would in a real library without the given extension. But the
implementation for this would be a lot more complicated and didn't really feel
worth it to me.